Moved all default settings logic to the default module.
